consolidate crowdin utils
